+++ sound-mind-studio/tests/test_image_scale_picker_dialog.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <QObject>
+
+class ImageScalePickerDialogTest : public QObject {
+    Q_OBJECT
+
+private slots:
+    void defaultsToRescaleToFitProject();
+    void selectingScaleVerticalKeepHorizontalUpdatesSelectedMode();
+    void selectingScaleHorizontalKeepVerticalUpdatesSelectedMode();
+    void selectingScaleVerticalProportionalUpdatesSelectedMode();
+    void selectingKeepNativeResolutionUpdatesSelectedMode();
+};